all: Add AU_915_928 band definition
